Game of quantum coins
# Quantum Coins
This is the game of quantum coins. It begins with a coin initially showing Heads. First the computer gets to play and has the choice to flip or not to flip the coin. Next you play and decide to flip or not flip the coin. At last the computers plays again and chooses to flip or not to flip the coin.  Under any conditions your moves are not revealed to the computer, and computers moves are not revealed to you. At the end, the state of the coin is revealed. If it is Tales you win, else the computer will win. 

The game is fair. For a classical computer you have a 50% chance of winning. In case of a Quantum Computer the computer wins every time. You can have a look at the Out0 algorithm, as it is used in this game.  The 'qasm_simulator' is used as the backend, as a ideal quantum computer(without any noise). To run it on a real quantum computer you need to load your account.

The magic actually happens in Part 2 of the program where the Out0 algorithm is used. No matter what you choose the computer always wins.